feat: complete testing mlops_team/Dockerfile.model_training
