+++ mlops_team/docs/weather_1hour_feature_list.md
@@ -1,0 +1,51 @@
+Weather 1hour Feature List
+
+초단기실황 (getUltraSrtNcst)
+→ 지금 이 순간 측정된 관측값 데이터 (머신러닝 입력값으로 적합)
+
+ObservationTime : 관측시각
+StationID : 지점번호
+WindDirection : 풍향
+WindSpeed : 풍속
+GustDirection : 돌풍방향
+GustSpeed : 돌풍속도
+GustTime : 돌풍시각
+LocalPressure : 현지기압
+SeaLevelPressure : 해면기압
+PressureTrend : 기압경향
+PressureChange : 기압변화량
+Temperature : 기온
+DewPointTemperature : 이슬점온도
+RelativeHumidity : 상대습도
+VaporPressure : 수증기압
+HourlyRainfall : 시간강수량
+DailyRainfall : 일강수량
+CumulativeRainfall : 누적강수량
+RainfallIntensity : 강수강도
+SnowDepth3Hr : 3시간적설
+DailySnowDepth : 일적설
+TotalSnowDepth : 총적설
+CurrentWeatherCode : 현재날씨코드
+PastWeatherCode : 과거날씨코드
+WeatherCode : 일기코드
+TotalCloudCover : 전운량
+MidLowCloudCover : 중하층운량
+LowestCloudHeight : 최저운고
+CloudType : 운형
+UpperCloudType : 상층운형
+MidCloudType : 중층운형
+LowCloudType : 하층운형
+Visibility : 시정
+SunshineDuration : 일조시간
+SolarRadiation : 일사량
+GroundCondition : 지면상태
+GroundTemperature : 지면온도
+SoilTemperature5cm : 5cm지중온도
+SoilTemperature10cm : 10cm지중온도
+SoilTemperature20cm : 20cm지중온도
+SoilTemperature30cm : 30cm지중온도
+SeaCondition : 해면상태
+WaveHeight : 파고
+MaxWindForce : 최대풍력
+PrecipitationData : 강수자료
+ObservationType : 관측유형
