+++ mlops_team/requirements.txt
@@ -2,8 +2,6 @@ fastapi
 uvicorn[standard]
 pandas
 scikit-learn
-
-# S3 사용을 위해 추가한 패키지들 (5월30일 추가함)
 python-dotenv 
 s3fs 
 pyarrow 
@@ -13,3 +11,4 @@ catboost
 numpy
 mlflow
 scipy
+boto3
