+++ scripts/evaluation_results.csv
@@ -1,0 +1,4 @@
+Model,RMSE,Training_Time
+RandomForest,1.0242879113570165,0.2697720527648926
+ExtraTrees,1.0727098796578212,0.1663529872894287
+XGBoost,1.184964192481141,0.14316797256469727
